+++ tests/timers.test.js
@@ -16,18 +16,18 @@ describe('timers test', () => {
 
     const timer1 = IR.SetTimeout(1000, mock1);
     const timer2 = IR.SetTimeout(2000, mock2);
-    expect(IR.irTimers.indexOf(timer1)).not.toBe(-1);
-    expect(IR.irTimers.indexOf(timer2)).not.toBe(-1);
+    expect(IR.mockTimers.indexOf(timer1)).not.toBe(-1);
+    expect(IR.mockTimers.indexOf(timer2)).not.toBe(-1);
     clock.tick(1000);
     expect(mock1).toBeCalled();
     expect(mock2).not.toBeCalled();
-    expect(IR.irTimers.indexOf(timer1)).toBe(-1);
-    expect(IR.irTimers.indexOf(timer2)).not.toBe(-1);
+    expect(IR.mockTimers.indexOf(timer1)).toBe(-1);
+    expect(IR.mockTimers.indexOf(timer2)).not.toBe(-1);
     clock.tick(1000);
     expect(mock1).toBeCalled();
     expect(mock2).toBeCalled();
-    expect(IR.irTimers.indexOf(timer1)).toBe(-1);
-    expect(IR.irTimers.indexOf(timer2)).toBe(-1);
+    expect(IR.mockTimers.indexOf(timer1)).toBe(-1);
+    expect(IR.mockTimers.indexOf(timer2)).toBe(-1);
   });
 
   it('IR.SetInterval', () => {
@@ -36,10 +36,10 @@ describe('timers test', () => {
     const IR = new IridiumMock();
 
     const timer = IR.SetInterval(1000, mock);
-    expect(IR.irTimers.indexOf(timer)).not.toBe(-1);
+    expect(IR.mockTimers.indexOf(timer)).not.toBe(-1);
     clock.tick(10000);
     expect(mock).toBeCalledTimes(10);
-    expect(IR.irTimers.indexOf(timer)).not.toBe(-1);
+    expect(IR.mockTimers.indexOf(timer)).not.toBe(-1);
   });
 
   it('IR.ClearTimeout', () => {
@@ -48,14 +48,14 @@ describe('timers test', () => {
     const IR = new IridiumMock();
 
     const timer = IR.SetTimeout(1000, mock);
-    expect(IR.irTimers.indexOf(timer)).not.toBe(-1);
+    expect(IR.mockTimers.indexOf(timer)).not.toBe(-1);
     clock.tick(500);
     expect(mock).not.toBeCalled();
-    expect(IR.irTimers.indexOf(timer)).not.toBe(-1);
+    expect(IR.mockTimers.indexOf(timer)).not.toBe(-1);
     IR.ClearTimeout(timer);
     clock.tick(1000);
     expect(mock).not.toBeCalled();
-    expect(IR.irTimers.indexOf(timer)).toBe(-1);
+    expect(IR.mockTimers.indexOf(timer)).toBe(-1);
   });
 
   it('IR.ClearInterval', () => {
@@ -64,13 +64,13 @@ describe('timers test', () => {
     const IR = new IridiumMock();
 
     const timer = IR.SetInterval(1000, mock);
-    expect(IR.irTimers.indexOf(timer)).not.toBe(-1);
+    expect(IR.mockTimers.indexOf(timer)).not.toBe(-1);
     clock.tick(500);
     expect(mock).not.toBeCalled();
-    expect(IR.irTimers.indexOf(timer)).not.toBe(-1);
+    expect(IR.mockTimers.indexOf(timer)).not.toBe(-1);
     IR.ClearInterval(timer);
     clock.tick(1000);
     expect(mock).not.toBeCalled();
-    expect(IR.irTimers.indexOf(timer)).toBe(-1);
+    expect(IR.mockTimers.indexOf(timer)).toBe(-1);
   });
 });