+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2020-05-10
 - IR.CreateDevice - added
+- IR.mockAppStart should be called to call IR.EVENT_START and connect all added devices
 - added jest tests (timers, variables, devices)
 
 ## 2020-01-15